+++ Tests/StencilTests/EnvironmentSpec.swift
@@ -320,7 +320,7 @@ final class EnvironmentBaseAndChildTemplateTests: XCTestCase {
     baseTemplate = try environment.loadTemplate(name: "invalid-base.html")
 
     try expectError(reason: "filter error",
-                    childToken: "block.super",
+                    childToken: "extends \"invalid-base.html\"",
                     baseToken: "target|unknown")
   }